refactor project, add a layout folder
