--- src/components/SideContent.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { Box } from "grommet";
-
-const SideContent = props => (
-  <Box fill background="light-2" align="center" justify="center" {...props} />
-);
-
-export default SideContent;
